update failing tests
