+++ test/asset-html/asset_html_script_importmap_inline_build_systemjs/asset_html_script_importmap_inline_build_systemjs.test.js
@@ -1,17 +1,23 @@
-import { basename } from "path"
 import { assert } from "@jsenv/assert"
-import { resolveDirectoryUrl, urlToRelativeUrl, readFile, resolveUrl } from "@jsenv/filesystem"
+import {
+  resolveDirectoryUrl,
+  urlToRelativeUrl,
+  readFile,
+  resolveUrl,
+  urlToBasename,
+} from "@jsenv/filesystem"
+
+import { buildProject } from "@jsenv/core"
 import { jsenvCoreDirectoryUrl } from "@jsenv/core/src/internal/jsenvCoreDirectoryUrl.js"
 import {
   findNodeByTagName,
-  getHtmlNodeAttributeByName,
+  getHtmlNodeTextNode,
 } from "@jsenv/core/src/internal/compiling/compileHtml.js"
 import { GENERATE_SYSTEMJS_BUILD_TEST_PARAMS } from "@jsenv/core/test/TEST_PARAMS_BUILD_SYSTEMJS.js"
-import { buildProject } from "@jsenv/core"
 
 const testDirectoryUrl = resolveDirectoryUrl("./", import.meta.url)
 const testDirectoryRelativeUrl = urlToRelativeUrl(testDirectoryUrl, jsenvCoreDirectoryUrl)
-const testDirectoryname = basename(testDirectoryRelativeUrl)
+const testDirectoryname = urlToBasename(testDirectoryRelativeUrl)
 const jsenvDirectoryRelativeUrl = `${testDirectoryRelativeUrl}.jsenv/`
 const buildDirectoryRelativeUrl = `${testDirectoryRelativeUrl}dist/systemjs/`
 const mainFilename = `${testDirectoryname}.html`
@@ -19,7 +25,7 @@ const entryPointMap = {
   [`./${testDirectoryRelativeUrl}${mainFilename}`]: "./main.html",
 }
 
-const { buildMappings } = await buildProject({
+await buildProject({
   ...GENERATE_SYSTEMJS_BUILD_TEST_PARAMS,
   // logLevel: "info",
   jsenvDirectoryRelativeUrl,
@@ -27,35 +33,17 @@ const { buildMappings } = await buildProject({
   entryPointMap,
 })
 
-const getBuildRelativeUrl = (urlRelativeToTestDirectory) => {
-  const relativeUrl = `${testDirectoryRelativeUrl}${urlRelativeToTestDirectory}`
-  const buildRelativeUrl = buildMappings[relativeUrl]
-  return buildRelativeUrl
-}
-
 const buildDirectoryUrl = resolveUrl(buildDirectoryRelativeUrl, jsenvCoreDirectoryUrl)
 const htmlBuildUrl = resolveUrl("main.html", buildDirectoryUrl)
 const htmlString = await readFile(htmlBuildUrl)
 const importmapScriptNode = findNodeByTagName(htmlString, "script")
-const importmapBuildRelativeUrl = getBuildRelativeUrl("import-map.importmap")
-const importmapBuildUrl = resolveUrl(importmapBuildRelativeUrl, buildDirectoryUrl)
 
-// ensure src is properly updated
+// ensure text content is correct
 {
-  const srcAttribute = getHtmlNodeAttributeByName(importmapScriptNode, "src")
-  const actual = srcAttribute.value
-  const expected = importmapBuildRelativeUrl
-  assert({ actual, expected })
-}
-
-// ensure importmap file content
-{
-  const importmapString = await readFile(importmapBuildUrl)
-  const actual = JSON.parse(importmapString)
+  const textNode = getHtmlNodeTextNode(importmapScriptNode)
+  const actual = JSON.parse(textNode.value)
   const expected = {
-    imports: {
-      foo: "./bar.js",
-    },
+    imports: {},
   }
   assert({ actual, expected })
 }
